requirements.txt.txt
fastapi==0.109.0
python-docx==1.0.1
openai==1.12.0
python-multipart==0.0.6
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
# Add these to existing requirements
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6